feat: add Python course lesson 2.5 (REPL) and improve table styling
- Add lesson 2.5 covering Python REPL usage (ES/EN)
- Fix schema to allow decimal order values (2.5, 3.5, etc.)
- Improve markdown table styling with full-width bleed effect
- Add white text for table headers (contrast on dark/light themes)
- Enable text wrapping in tables for mobile readability
- Apply responsive table styles across tutorials and posts

+++ src/schemas/blog.ts
@@ -116,7 +116,7 @@ export const tutorialsSchema = z.object({
 
   // Optional metadata
   course: z.string().optional(), // reference to courses collection
-  order: z.number().int().positive().optional(), // Tutorial order within course (1, 2, 3...)
+  order: z.number().positive().optional(), // Tutorial order within course (1, 2, 2.5, 3...)
   cover: z.string().optional(), // Relative path to cover image
   update_date: z.coerce.date().optional(), // When the tutorial was last updated
